+++ CHANGELOG.md
@@ -4,8 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]## [0.1.0.0b6]
+## [Unreleased]## [0.1.1.0b0]
+### Changed
+- Version bump-up to 0.1.1.0 due to many new features
 ### Added
+- Home page scenario duplicate, rename, delete
+- DoDashApp.db_echo flag for DB connection debugging
 
 ## [0.1.0.0b5] - 2022-01-05
 ### Changed